requirements.txt
Flask==2.3.3
psycopg==3.1.9
google-genai==0.3.0
gunicorn==21.2.0
